LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-center-offset-change-expected.txt


FAIL Anchored initially have the same width as the anchor assert_equals: expected 0 but got 100
FAIL Increase the height of the anchor to move the anchor-center offset assert_equals: expected 50 but got 200
